renamed instances of require("gypsy... to require("chat-gypsy...
